Bring back unzipper to proxy
